+++ packages/visual-editor/src/components/migrations/migrationRegistry.ts
@@ -4,6 +4,10 @@ import { addHeadingAlignmentMigration } from "./0002_add_heading_alignment.ts";
 import { adjustStructFields } from "./0003_adjust_struct_fields.ts";
 import { addDirectoryRootPropMigration } from "./0004_add_directory_root_prop.ts";
 import { addPromoHeadingStylesMigration } from "./0005_add_promo_heading_styles.ts";
+import { updateImageStylingMigration } from "./0006_update_image_styling.ts";
+import { addCardStylesMigration } from "./0007_add_card_styles.ts";
+import { addBreadcrumbsDirectoryBackgroundMigration } from "./0008_add_breadcrumbs_directory_background.ts";
+import { addShowAverageReviewMigration } from "./0009_add_show_average_reviews.ts";
 
 // To add a migration:
 // Create a new file in this directory that exports a Migration
@@ -16,4 +20,8 @@ export const migrationRegistry: MigrationRegistry = [
   adjustStructFields,
   addDirectoryRootPropMigration,
   addPromoHeadingStylesMigration,
+  updateImageStylingMigration,
+  addCardStylesMigration,
+  addBreadcrumbsDirectoryBackgroundMigration,
+  addShowAverageReviewMigration,
 ];