+++ packages/visual-editor/src/components/pageSections/CoreInfoSection.tsx
@@ -194,7 +194,7 @@ const coreInfoSectionFields: Fields<CoreInfoSectionProps> = {
       heading: YextField(msg("fields.heading", "Heading"), {
         type: "object",
         objectFields: {
-          level: YextField(msg("fields.headingLevel", "Level"), {
+          level: YextField(msg("fields.level", "Level"), {
             type: "select",
             hasSearch: true,
             options: "HEADING_LEVEL",