+++ packages/visual-editor/src/components/pageSections/FAQsSection.test.tsx
@@ -3,7 +3,8 @@ import { describe, it, expect } from "vitest";
 import {
   axe,
   ComponentTest,
-  viewports,
+  delay,
+  transformTests,
 } from "../testing/componentTests.setup.ts";
 import { render as reactRender, act } from "@testing-library/react";
 import {
@@ -38,18 +39,12 @@ const tests: ComponentTest[] = [
     document: {},
     props: { ...FAQSection.defaultProps },
     version: migrationRegistry.length,
-    tests: async (page) => {
-      expect(page.getByText("Frequently Asked Questions")).toBeVisible();
-    },
   },
   {
     name: "default props with document data",
     document: { c_faq: faqData },
     props: { ...FAQSection.defaultProps },
     version: migrationRegistry.length,
-    tests: async (page) => {
-      expect(page.getByText("Frequently Asked Questions")).toBeVisible();
-    },
   },
   {
     name: "version 0 props with entity values",
@@ -79,20 +74,14 @@ const tests: ComponentTest[] = [
       liveVisibility: true,
     },
     version: 0,
-    tests: async (page) => {
+    interactions: async (page) => {
       const q1 = page.getByText("What services do you offer?");
       const q2 = page.getByText("Do you have parking available?");
-      expect(page.getByText("test name")).toBeVisible();
-      expect(q1).toBeVisible();
-      expect(q2).toBeVisible();
       await act(async () => {
         await q1.click();
-      });
-      expect(page.getByText("delivery")).toBeVisible();
-      await act(async () => {
         await q2.click();
+        await delay(200);
       });
-      expect(page.getByText("limited")).toBeVisible();
     },
   },
   {
@@ -127,27 +116,16 @@ const tests: ComponentTest[] = [
       liveVisibility: true,
     },
     version: 0,
-    tests: async (page) => {
+    interactions: async (page) => {
       const q1 = page.getByText("Test Question 1");
       const q2 = page.getByText("Test Question 2");
-      expect(page.getByText("Frequently Asked Questions")).toBeVisible();
-      expect(q1).toBeVisible();
-      expect(q2).toBeVisible();
       await act(async () => {
         await q1.click();
-      });
-      expect(page.getByText("Answer 1")).toBeVisible();
-      await act(async () => {
         await q2.click();
+        await delay(200);
       });
-      expect(page.getByText("Answer 2")).toBeVisible();
     },
   },
-];
-
-const testsWithViewports: ComponentTest[] = [
-  ...tests.map((t) => ({ ...t, viewport: viewports[0] })),
-  ...tests.map((t) => ({ ...t, viewport: viewports[1] })),
 ];
 
 describe("FAQSection", async () => {
@@ -159,14 +137,15 @@ describe("FAQSection", async () => {
       },
     },
   };
-  it.each(testsWithViewports)(
-    "renders $name $viewport.name",
+  it.each(transformTests(tests))(
+    "$viewport.name $name",
     async ({
       document,
+      name,
       props,
-      tests,
+      interactions,
       version,
-      viewport: { width, height } = viewports[0],
+      viewport: { width, height, name: viewportName },
     }) => {
       const data = migrate(
         {
@@ -185,6 +164,7 @@ describe("FAQSection", async () => {
         migrationRegistry,
         puckConfig
       );
+
       const { container } = reactRender(
         <VisualEditorProvider templateProps={{ document }}>
           <Render config={puckConfig} data={data} />
@@ -192,10 +172,20 @@ describe("FAQSection", async () => {
       );
 
       await page.viewport(width, height);
-      await page.screenshot();
+      await delay(600);
+
+      await expect(`FAQsSection/[${viewportName}] ${name}`).toMatchScreenshot();
       const results = await axe(container);
       expect(results).toHaveNoViolations();
-      await tests(page);
+
+      if (interactions) {
+        await interactions(page);
+        await expect(
+          `FAQsSection/[${viewportName}] ${name} (after interactions)`
+        ).toMatchScreenshot();
+        const results = await axe(container);
+        expect(results).toHaveNoViolations();
+      }
     }
   );
 });