+++ packages/visual-editor/src/components/pageSections/ProductSection.test.tsx
@@ -3,7 +3,8 @@ import { describe, it, expect } from "vitest";
 import {
   axe,
   ComponentTest,
-  viewports,
+  transformTests,
+  delay,
 } from "../testing/componentTests.setup.ts";
 import { render as reactRender } from "@testing-library/react";
 import {
@@ -25,7 +26,7 @@ const productsData = {
         linkType: "OTHER",
       },
       description: {
-        html: '\u003ch1 dir="ltr" style="font-size: 14.67px; font-weight: 400; line-height: 18.67px; color: rgb(0, 0, 0); margin: 0; padding: 3px 2px 3px 2px; position: relative;"\u003e\u003cspan\u003eOur signature burger! A juicy beef patty topped with melted cheese, crisp lettuce, ripe tomato, and our special Galaxy sauce, all served on a toasted sesame seed bun.\u003c/span\u003e\u003c/p\u003e',
+        html: '<p dir="ltr" style="font-size: 14.67px; font-weight: 400; line-height: 18.67px; color: rgb(0, 0, 0); margin: 0; padding: 3px 2px 3px 2px; position: relative;"><span>Our signature burger! A juicy beef patty topped with melted cheese, crisp lettuce, ripe tomato, and our special Galaxy sauce, all served on a toasted sesame seed bun.</span></p>',
       },
       image: {
         height: 2048,
@@ -117,24 +118,12 @@ const tests: ComponentTest[] = [
     document: {},
     props: { ...ProductSection.defaultProps },
     version: migrationRegistry.length,
-    tests: async (page) => {
-      expect(page.getByText("Featured Products")).toBeVisible();
-      expect(document.body.textContent).toContain(
-        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
-      );
-    },
   },
   {
     name: "default props with document data",
     document: { c_products: productsData },
     props: { ...ProductSection.defaultProps },
     version: migrationRegistry.length,
-    tests: async (page) => {
-      expect(page.getByText("Featured Products")).toBeVisible();
-      expect(document.body.textContent).toContain(
-        "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
-      );
-    },
   },
   {
     name: "version 0 props with entity values",
@@ -165,16 +154,6 @@ const tests: ComponentTest[] = [
       liveVisibility: true,
     },
     version: 0,
-    tests: async (page) => {
-      expect(page.getByText("Test Name")).toBeVisible();
-      expect(page.getByText("Galaxy Burger")).toBeVisible();
-      expect(page.getByText("Burgers")).toBeVisible();
-      expect(page.getByText("Order Now").elements()).toHaveLength(2);
-      expect(page.getByText("Galaxy Salad")).toBeVisible();
-      expect(page.getByText("Galaxy Milkshake")).toBeVisible();
-      expect(page.getByText("Desserts")).toBeVisible();
-      expect(page.getByText("cherry")).toBeVisible();
-    },
   },
   {
     name: "version 0 props with constant value",
@@ -215,19 +194,92 @@ const tests: ComponentTest[] = [
       liveVisibility: true,
     },
     version: 0,
-    tests: async (page) => {
-      expect(page.getByText("Featured Products")).toBeVisible();
-      expect(page.getByText("Product 1")).toBeVisible();
-      expect(page.getByText("Category")).toBeVisible();
-      expect(page.getByText("Description")).toBeVisible();
-      expect(page.getByText("CTA")).toBeVisible();
-    },
   },
-];
-
-const testsWithViewports: ComponentTest[] = [
-  ...tests.map((t) => ({ ...t, viewport: viewports[0] })),
-  ...tests.map((t) => ({ ...t, viewport: viewports[1] })),
+  {
+    name: "version 7 props with entity values",
+    document: { c_products: productsData, name: "Test Name" },
+    props: {
+      data: {
+        heading: {
+          field: "name",
+          constantValue: "Featured Products",
+          constantValueEnabled: false,
+          constantValueOverride: {},
+        },
+        products: {
+          field: "c_products",
+          constantValue: { products: [] },
+          constantValueEnabled: false,
+          constantValueOverride: {},
+        },
+      },
+      styles: {
+        backgroundColor: {
+          bgColor: "bg-palette-secondary-light",
+          textColor: "text-black",
+        },
+        heading: {
+          level: 2,
+          align: "left",
+        },
+        cards: {
+          backgroundColor: {
+            bgColor: "bg-palette-primary-light",
+            textColor: "text-black",
+          },
+          headingLevel: 4,
+        },
+      },
+      liveVisibility: true,
+    },
+    version: 7,
+  },
+  {
+    name: "version 7 props with constant value",
+    document: { c_products: productsData },
+    props: {
+      data: {
+        heading: {
+          field: "name",
+          constantValue: "Featured Products",
+          constantValueEnabled: true,
+        },
+        products: {
+          field: "c_products",
+          constantValue: {
+            products: [
+              {
+                name: "Product 1",
+                category: "Category",
+                description: "Description",
+                cta: { label: "CTA" },
+              },
+            ],
+          },
+          constantValueEnabled: true,
+        },
+      },
+      styles: {
+        backgroundColor: {
+          bgColor: "bg-palette-secondary-light",
+          textColor: "text-black",
+        },
+        heading: {
+          level: 2,
+          align: "left",
+        },
+        cards: {
+          backgroundColor: {
+            bgColor: "bg-palette-primary-light",
+            textColor: "text-black",
+          },
+          headingLevel: 3,
+        },
+      },
+      liveVisibility: true,
+    },
+    version: 7,
+  },
 ];
 
 describe("ProductSection", async () => {
@@ -239,14 +291,15 @@ describe("ProductSection", async () => {
       },
     },
   };
-  it.each(testsWithViewports)(
-    "renders $name $viewport.name",
+  it.each(transformTests(tests))(
+    "$viewport.name $name",
     async ({
       document,
+      name,
       props,
-      tests,
+      interactions,
       version,
-      viewport: { width, height } = viewports[0],
+      viewport: { width, height, name: viewportName },
     }) => {
       const data = migrate(
         {
@@ -265,6 +318,7 @@ describe("ProductSection", async () => {
         migrationRegistry,
         puckConfig
       );
+
       const { container } = reactRender(
         <VisualEditorProvider templateProps={{ document }}>
           <Render config={puckConfig} data={data} />
@@ -272,11 +326,22 @@ describe("ProductSection", async () => {
       );
 
       await page.viewport(width, height);
-      await page.screenshot();
+      await delay(600);
 
+      await expect(
+        `ProductSection/[${viewportName}] ${name}`
+      ).toMatchScreenshot();
       const results = await axe(container);
       expect(results).toHaveNoViolations();
-      await tests(page);
+
+      if (interactions) {
+        await interactions(page);
+        await expect(
+          `ProductSection/[${viewportName}] ${name} (after interactions)`
+        ).toMatchScreenshot();
+        const results = await axe(container);
+        expect(results).toHaveNoViolations();
+      }
     }
   );
 });