+++ packages/visual-editor/src/components/testing/componentTests.d.ts
@@ -1,0 +1,18 @@
+import "vitest";
+
+interface CustomMatchers<R = unknown> {
+  toMatchScreenshot: () => Promise<R>;
+}
+
+declare module "vitest" {
+  interface Matchers<T = any> extends CustomMatchers<T> {}
+}
+
+declare module "@vitest/browser/context" {
+  interface BrowserCommands {
+    compareScreenshot: (
+      screenshotName: string,
+      updatedScreenshotData: string
+    ) => Promise<number>;
+  }
+}